pyproject.toml
[project]
name = "elderlycentrecrm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "faker>=40.4.0",
    "fastapi>=0.129.2",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.41.0",
]
